requirements.txt
pyserial
numpy
scipy